Miscellaneous fixes

+++ db/migrations/000001_initialise_schema.up.sql
@@ -122,7 +122,7 @@ CREATE TYPE ingredient_unit_enum AS ENUM (
 CREATE TABLE recipe_ingredients (
   recipe_id SERIAL NOT NULL REFERENCES recipes(id) ON DELETE CASCADE,
   ingredient_id SERIAL NOT NULL REFERENCES ingredients(id) ON DELETE CASCADE,
-  quantity DECIMAL(10, 2) NOT NULL,
+  quantity FLOAT NOT NULL,
   unit ingredient_unit_enum NOT NULL,
   created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
   updated_at TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
